A few more places where "unsigned long" should have been used.

+++ generic/tkCanvas.c
@@ -4843,7 +4843,7 @@ PickCurrentItem(
 				 * ButtonRelease, or MotionNotify. */
 {
     double coords[2];
-    unsigned int buttonDown;
+    unsigned long buttonDown;
     Tk_Item *prevItemPtr;
 #ifndef USE_OLD_TAG_SEARCH
     SearchUids *searchUids = GetStaticUids();
